# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: M360 API
  version: 1.0.0
  description: |
    M360 SME Lending CRM API for Shariah-compliant Murabaha financing in Saudi Arabia.
    Bilingual Arabic/English with full loan lifecycle management, Saudi government
    integrations (SIMAH, Bayan, Nafath, Yaqeen, Watheq), and role-based access control.

    ## Roles
    - `admin` - Full system access
    - `manager` - Management-level access
    - `loan_officer` - Loan origination and management
    - `credit_analyst` - Credit assessment and committee packages
    - `compliance_officer` - Compliance oversight
    - `collections_officer` - Collections management
    - `data_entry` - Data entry operations

    ## Application Workflow
    Draft -> Submitted -> PreApproved -> DocsCollected -> CreditAssessment -> CommitteeReview -> Approved/Rejected -> Disbursed
  contact:
    name: M360 Team
  license:
    name: Proprietary
servers:
  - url: /api/v1
    description: API v1 base path
security:
  - BearerAuth: []
...
paths:
...
  /contacts:
    post:
      tags: [Contacts]
      summary: Create a contact
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ContactCreateRequest'
      responses:
        '201':
          description: Contact created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Contact'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /contacts/{id}:
    get:
      tags: [Contacts]
      summary: Get contact by ID
      parameters:
        - $ref: '#/components/parameters/IDParam'
      responses:
        '200':
          description: Contact details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Contact'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Contacts]
      summary: Update a contact
      parameters:
        - $ref: '#/components/parameters/IDParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ContactUpdateRequest'
      responses:
        '200':
          description: Contact updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Contact'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
    delete:
      tags: [Contacts]
      summary: Delete a contact
      description: Requires admin or manager role.
      parameters:
        - $ref: '#/components/parameters/IDParam'
      responses:
        '204':
          description: Contact deleted
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
  /contacts/organization/{org_id}:
    get:
      tags: [Contacts]
      summary: List contacts by organization
      parameters:
        - name: org_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: List of contacts for the organization
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Contact'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'

  # ──────────────────────────────────────────────
  # Leads
  # ──────────────────────────────────────────────
...
  /packages:
    post:
      tags: [Committee]
      summary: Create a committee package
      description: Requires credit_analyst, admin, or manager role.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CommitteePackageCreateRequest'
      responses:
        '201':
          description: Committee package created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommitteePackage'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
  /packages/{id}:
    get:
      tags: [Committee]
      summary: Get committee package by ID
      parameters:
        - $ref: '#/components/parameters/IDParam'
      responses:
        '200':
          description: Committee package details (includes votes)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommitteePackage'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /packages/application/{app_id}:
    get:
      tags: [Committee]
      summary: Get committee package by application ID
      parameters:
        - name: app_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Committee package for the application
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommitteePackage'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /packages/{id}/vote:
    post:
      tags: [Committee]
      summary: Cast a vote on a committee package
      description: Requires manager role.
      parameters:
        - $ref: '#/components/parameters/IDParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CastVoteRequest'
      responses:
        '200':
          description: Vote cast, returns updated package
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommitteePackage'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'

  # ──────────────────────────────────────────────
  # Facilities
  # ──────────────────────────────────────────────
...
  /collections:
    post:
      tags: [Collections]
      summary: Create a collection action
      description: Requires collections_officer, admin, or manager role.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CollectionCreateRequest'
      responses:
        '201':
          description: Collection action created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CollectionAction'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
  /collections/facility/{id}:
    get:
      tags: [Collections]
      summary: List collection actions by facility
      parameters:
        - $ref: '#/components/parameters/IDParam'
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/OffsetParam'
      responses:
        '200':
          description: Collection actions for the facility
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/CollectionAction'
                  meta:
                    type: object
                    properties:
                      limit:
                        type: integer
                      offset:
                        type: integer
                      total:
                        type: integer
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /collections/overdue:
    get:
      tags: [Collections]
      summary: Get overdue facilities summary
      responses:
        '200':
          description: Overdue summary
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OverdueSummary'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'

  # ──────────────────────────────────────────────
  # Documents
  # ──────────────────────────────────────────────
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token obtained from /auth/login or /auth/refresh
  parameters:
    IDParam:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    LimitParam:
      name: limit
      in: query
      description: Number of items per page (default 20)
      schema:
        type: integer
        default: 20
    OffsetParam:
      name: offset
      in: query
      description: Number of items to skip
      schema:
        type: integer
        default: 0
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Unauthorized - missing or invalid JWT
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Forbidden - insufficient role
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
          description: Error message
...
    Contact:
      type: object
      properties:
        id:
          type: string
          format: uuid
        organization_id:
          type: string
          format: uuid
        name_en:
          type: string
        name_ar:
          type: string
        national_id:
          type: string
          nullable: true
        role:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        nafath_verified:
          type: boolean
        simah_score:
          type: integer
          nullable: true
        is_guarantor:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    ContactCreateRequest:
      type: object
      required: [organization_id, name_en]
      properties:
        organization_id:
          type: string
          format: uuid
        name_en:
          type: string
        name_ar:
          type: string
        national_id:
          type: string
        role:
          type: string
        phone:
          type: string
        email:
          type: string
        is_guarantor:
          type: boolean
    ContactUpdateRequest:
      type: object
      properties:
        name_en:
          type: string
        name_ar:
          type: string
        national_id:
          type: string
        role:
          type: string
        phone:
          type: string
        email:
          type: string
        is_guarantor:
          type: boolean

    # ── Lead ──
...
    CommitteeDecision:
      type: string
      enum: [pending, approved, rejected, approved_with_conditions]
    VoteDecision:
      type: string
      enum: [approve, reject, approve_with_conditions]
    CommitteeVote:
      type: object
      properties:
        id:
          type: string
          format: uuid
        package_id:
          type: string
          format: uuid
        voter_id:
          type: string
          format: uuid
        decision:
          $ref: '#/components/schemas/VoteDecision'
        comments:
          type: string
          nullable: true
        voted_at:
          type: string
          format: date-time
    CommitteePackage:
      type: object
      properties:
        id:
          type: string
          format: uuid
        application_id:
          type: string
          format: uuid
        prepared_by:
          type: string
          format: uuid
        risk_score:
          type: integer
          nullable: true
        recommendation:
          type: string
          nullable: true
        financial_analysis:
          type: object
          description: Arbitrary JSON
        decision:
          $ref: '#/components/schemas/CommitteeDecision'
        decision_date:
          type: string
          format: date-time
          nullable: true
        conditions:
          type: string
          nullable: true
        quorum_required:
          type: integer
        votes:
          type: array
          items:
            $ref: '#/components/schemas/CommitteeVote'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CommitteePackageCreateRequest:
      type: object
      required: [application_id, quorum_required]
      properties:
        application_id:
          type: string
          format: uuid
        risk_score:
          type: integer
        recommendation:
          type: string
        quorum_required:
          type: integer
    CastVoteRequest:
      type: object
      required: [decision]
      properties:
        decision:
          $ref: '#/components/schemas/VoteDecision'
        comments:
          type: string

    # ── Facility ──
...
    CollectionAction:
      type: object
      properties:
        id:
          type: string
          format: uuid
        facility_id:
          type: string
          format: uuid
        officer_id:
          type: string
          format: uuid
        action_type:
          type: string
          enum: [sms_reminder, phone_call, formal_notice, restructuring]
        description:
          type: string
          nullable: true
        next_action_date:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
    CollectionCreateRequest:
      type: object
      required: [facility_id, action_type]
      properties:
        facility_id:
          type: string
          format: uuid
        action_type:
          type: string
          enum: [sms_reminder, phone_call, formal_notice, restructuring]
        description:
          type: string
        next_action_date:
          type: string
          format: date-time
    OverdueFacility:
      type: object
      properties:
        id:
          type: string
          format: uuid
        facility_number:
          type: string
        borrower_name:
          type: string
        overdue_amount:
          type: number
          format: double
        days_overdue:
          type: integer
        last_collection_date:
          type: string
          format: date-time
          nullable: true
        collection_count:
          type: integer
    OverdueSummary:
      type: object
      properties:
        total_overdue_facilities:
          type: integer
        overdue_facilities:
          type: array
          items:
            $ref: '#/components/schemas/OverdueFacility'

    # ── Document ──
...